+++ src/features/calendar/components/views/FiveDayView/index.tsx
@@ -1,0 +1,2 @@
+export { FiveDayView } from './FiveDayView'
+export type { FiveDayViewProps } from './FiveDayView.types'